Added basic pressure-based hotspot code
When a pointer motion event is sent to a hotspot, it now also receives relative
unaccelerated motion data. When the cursor is within the hotspot's area and is
moving towards the corner, "pressure" is calculated and added. If it reaches the
value of 128.0 (determined subjectively by trying to match GNOME's behavior),
the hotspot's action is activated immediately.

This does not replace default timer-based behavior of hotspots, as while the new
behavior works well with mice (which can send motion beyond screen borders), the
same can not be said for pointing devices (like graphical tablets) or
touchscreens.